requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.29.0
accelerate>=0.20.3
huggingface_hub>=0.15.1
streamlit>=1.18.1
Pillow>=9.0.0
